requirements.txt
streamlit
pydantic
python-dotenv
pyyaml
chromadb
httpx
openai
